+++ fhir_db/lib/r5/resource_dao.dart
@@ -65,7 +65,7 @@ class ResourceDao {
       if (resource.resourceType != null) {
         await _addResourceType(password, resource.resourceType!);
 
-        _setStoreType(resource.resourceTypeString()!);
+        _setStoreType(resource.resourceTypeString!);
 
         return resource.id == null
             ? await _insert(password, resource)