+++ fhir/CHANGELOG.md
@@ -170,7 +170,7 @@
 
 ## [0.1.3]
 
-* added a resourceTypeString() method to easily get string from enum (needed for Db)
+* added a resourceTypeString method to easily get string from enum (needed for Db)
 
 ## [0.1.2]